requirements.txt
python-dotenv
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-cloud-vision>=3.4.0
PyYAML>=6.0
Pillow>=10.0.0
tqdm>=4.65.0
torch>=2.0.0
transformers>=4.36.0
sentencepiece>=0.1.99
peft>=0.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
qdrant-client>=1.7.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.7
google-generativeai>=0.3.0
openai-whisper>=20231117
beautifulsoup4>=4.12.0
requests>=2.31.0

langdetect>=1.0.9
transformers>=4.38.0
accelerate>=0.26.0
sentence-transformers>=2.2.2
FlagEmbedding>=1.2.0
